requirements.txt
discord-py-interactions
aiohttp
requests
beautifulsoup4